--- a/comp/netflow/flowaggregator/aggregator.go
+++ b/comp/netflow/flowaggregator/aggregator.go
@@ -137,7 +137,7 @@ func (agg *FlowAggregator) run() {
 func (agg *FlowAggregator) sendFlows(flows []*common.Flow, flushTime time.Time) {
 	for _, flow := range flows {
 		flowPayload := buildPayload(flow, agg.hostname, flushTime)
-		payloadBytes, err := json.Marshal(flowPayload)
+		payloadBytes, err := flowPayload.MarshalJSON()
 		if err != nil {
 			agg.logger.Errorf("Error marshalling device metadata: %s", err)
 			continue