--- a/comp/netflow/goflowlib/flowstate.go
+++ b/comp/netflow/goflowlib/flowstate.go
@@ -10,6 +10,7 @@ import (
 	"fmt"
 	"github.com/DataDog/datadog-agent/comp/netflow/config"
 	"github.com/DataDog/datadog-agent/comp/netflow/goflowlib/netflowstate"
+	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
 
 	"github.com/netsampler/goflow2/decoders/netflow/templates"
 	"go.uber.org/atomic"
@@ -52,6 +53,7 @@ func StartFlowRoutine(
 	fieldMappings []config.Mapping,
 	flowInChan chan *common.Flow,
 	logger log.Component,
+	sender sender.Sender,
 	atomicErr *atomic.String,
 	listenerFlowCount *atomic.Int64) (*FlowStateWrapper, error) {
 	var flowState FlowRunnableState
@@ -68,7 +70,7 @@ func StartFlowRoutine(
 		}
 		defer templateSystem.Close(ctx)
 
-		state := netflowstate.NewStateNetFlow(fieldMappings)
+		state := netflowstate.NewStateNetFlow(fieldMappings, sender)
 		state.Format = formatDriver
 		state.Logger = logrusLogger
 		state.TemplateSystem = templateSystem