--- a/comp/netflow/goflowlib/formatdriver.go
+++ b/comp/netflow/goflowlib/formatdriver.go
@@ -42,11 +42,16 @@ func (d *AggregatorFormatDriver) Init(context.Context) error {
 
 // Format desc
 func (d *AggregatorFormatDriver) Format(data interface{}) ([]byte, []byte, error) {
-	flow, ok := data.(*flowpb.FlowMessage)
-	if !ok {
+	switch flow := data.(type) {
+	case *flowpb.FlowMessage:
+		d.listenerFlowCount.Add(1)
+		d.flowAggIn <- ConvertFlow(flow, d.namespace)
+	case *common.FlowMessageWithAdditionalFields:
+		d.listenerFlowCount.Add(1)
+		d.flowAggIn <- ConvertFlowWithCustomFields(flow, d.namespace)
+	default:
 		return nil, nil, fmt.Errorf("message is not flowpb.FlowMessage")
 	}
-	d.listenerFlowCount.Add(1)
-	d.flowAggIn <- ConvertFlow(flow, d.namespace)
+
 	return nil, nil, nil
 }