--- a/comp/netflow/goflowlib/formatdriver.go
+++ b/comp/netflow/goflowlib/formatdriver.go
@@ -48,7 +48,7 @@ func (d *AggregatorFormatDriver) Format(data interface{}) ([]byte, []byte, error
 		d.flowAggIn <- ConvertFlow(flow, d.namespace)
 	case *common.FlowMessageWithAdditionalFields:
 		d.listenerFlowCount.Add(1)
-		d.flowAggIn <- ConvertFlowWithCustomFields(flow, d.namespace)
+		d.flowAggIn <- ConvertFlowWithAdditionalFields(flow, d.namespace)
 	default:
 		return nil, nil, fmt.Errorf("message is not flowpb.FlowMessage or common.FlowMessageWithAdditionalFields")
 	}